day2 task21&task22 solution

+++ src/main/java/pl/java/exercises/day2/task20_21_22/Hexagon.java
@@ -3,9 +3,9 @@ package pl.java.exercises.day2.task20_21_22;
 import java.util.Objects;
 
 public class Hexagon extends Shape{
-    private final int a;
+    private final float a;
 
-    public Hexagon(int a) {
+    public Hexagon(float a) {
         this.a = a;
     }
